backend/requirements.txt
fastapi==0.110.1 
uvicorn==0.25.0 
motor==3.3.1 
pymongo==4.5.0 
python-dotenv==1.2.1 
passlib[bcrypt]==1.7.4 
pydantic==2.12.5 
PyJWT==2.11.0 
python-multipart==0.0.22 
